requirements.txt
flask
transformers
scikit-learn
python-dotenv
torch
protobuf
tiktoken
flask-cors
datasets